move BaseDimensions to dimension_math
